requirements.txt
requests==2.32.3
pycryptodome==3.21.0
pyyaml==6.0.2
